reviewing old test that had an old syntax, were testing non-existing methods etc.; removing separate files for test with combiner and inner splitters; merging tests for a single taks from test_node.py and test_satraexample.py (removed tests who were testing the same parts of the code); leaving only workflow tests in test_satraexample.py (renaming to test_workflow); extracting old tests for interfaces from nipype1 to test_nipype1interfaces.py (will have to be updated)
